reorganized file directories ( separated frontend and backend )
