+++ frontend/src/components/Carousel.jsx
